working decoupled version
