+++ src/structbond/StructBondModule.jl
@@ -1,7 +1,10 @@
 module StructBondModule 
     import PlutoUI.Experimental: wrapped
 
+    export BondTable, StructBond, @NTBond, @BondsList, Popout, @popoutasfield,
+    @typeasfield, popoutwrap, @fieldbond, @fielddescription, @fielddata
+
     include("basics.jl")
-    include("main_definition.jl")
+    include("main_definitions.jl")
     include("macro.jl")
 end
